server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx"
    ,"python-dotenv"
]

[project.scripts]
tamos-server = "cli:main"
